Move ring buffer to its own package
